Adding EchoAnalyzer
