+++ src/alogamous/analyzer.py
@@ -16,8 +16,3 @@ def analyze_log_stream(analyzers, in_stream, out_stream):
             analyzer.read_log_line(line.rstrip())
     for analyzer in analyzers:
         analyzer.report(out_stream)
-
-
-if __name__ == "__main__":
-    analyze_log_stream(sys.stdin, sys.stdout)
-
