--- src/alogamous/example.py
@@ -1,2 +1,0 @@
-def hello_world():
-    return "Hello World"
